+++ src/documentation/sections/subSections/index.js
@@ -1,0 +1,8 @@
+import ColorsProp from './ColorsProp.vue';
+import TableLoaderProps from './TableLoaderProps.vue';
+
+
+export {
+	ColorsProp,
+	TableLoaderProps,
+};
